requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
python-dotenv
loguru
openpyxl
pandas
scikit-learn